goofy change to repurpose dismiss button for hard-coded Learn more link
Hard code a Learn more link for the banner message announcing availability of the issued-in-2021 about-2020 WRS ETF SoB. This is to deliver this Learn more link feature for this specific message in time for the message launch. Thereafter, the intention is to implement the "Learn more" link feature right.